NewVersionProperties&CorrectioDB&AddUser&AddPageSecurityConfiguration

+++ src/main/resources/db/migration/V1_4__init_cod_fish_bike_to_buy.sql
@@ -6,7 +6,7 @@ CREATE TABLE bike_to_buy
     serial          VARCHAR(32)    NOT NULL,
     brand           VARCHAR(32)    NOT NULL,
     model           VARCHAR(128)   NOT NULL,
-    production_year SMALLINT       NOT NULL,
+    production_year INT            NOT NULL,
     color           VARCHAR(32),
     price           NUMERIC(19, 2) NOT NULL,
     PRIMARY KEY (bike_to_buy_id),
